+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.6.0]
+### Added
+- Hunt! feature that let you search search all potential indicators present in the web page visible in the current active tab of the browser.
+
+### Changed
+- Behaviour of indicators text selection. Now only submitted indicator will be saved inside the input text field of the popup.
+- Buttons icons inside the input text field.
+
 ### Fixed
 - Behaviour of the domain extraction feature icon in the textfield